move SumAggregate.h to function/lib
