+++ velox/functions/prestosql/ArraySum.cpp
@@ -20,8 +20,8 @@
 #include "velox/expression/EvalCtx.h"
 #include "velox/expression/Expr.h"
 #include "velox/expression/VectorFunction.h"
+#include "velox/functions/lib/CheckedArithmetic.h"
 #include "velox/functions/lib/LambdaFunctionUtil.h"
-#include "velox/functions/prestosql/CheckedArithmetic.h"
 
 namespace facebook::velox::functions {
 namespace {
