+++ velox/functions/sparksql/aggregates/CMakeLists.txt
@@ -13,8 +13,13 @@
 # limitations under the License.
 add_library(
   velox_functions_spark_aggregates
-  BitwiseXorAggregate.cpp BloomFilterAggAggregate.cpp FirstLastAggregate.cpp
-  AverageAggregate.cpp MinMaxByAggregate.cpp Register.cpp)
+  AverageAggregate.cpp
+  BitwiseXorAggregate.cpp
+  BloomFilterAggAggregate.cpp
+  FirstLastAggregate.cpp
+  MinMaxByAggregate.cpp
+  Register.cpp
+  SumAggregate.cpp)
 
 target_link_libraries(velox_functions_spark_aggregates fmt::fmt velox_exec
                       velox_expression_functions velox_aggregates velox_vector)
