+++ velox/functions/sparksql/aggregates/Register.cpp
@@ -19,6 +19,7 @@
 #include "velox/functions/sparksql/aggregates/AverageAggregate.h"
 #include "velox/functions/sparksql/aggregates/BitwiseXorAggregate.h"
 #include "velox/functions/sparksql/aggregates/BloomFilterAggAggregate.h"
+#include "velox/functions/sparksql/aggregates/SumAggregate.h"
 
 namespace facebook::velox::functions::aggregate::sparksql {
 
@@ -31,5 +32,6 @@ void registerAggregateFunctions(const std::string& prefix) {
   registerBitwiseXorAggregate(prefix);
   registerBloomFilterAggAggregate(prefix + "bloom_filter_agg");
   registerAverage(prefix + "avg");
+  registerSum(prefix + "sum");
 }
 } // namespace facebook::velox::functions::aggregate::sparksql
