+++ velox/functions/sparksql/aggregates/SumAggregate.cpp
@@ -14,14 +14,11 @@
  * limitations under the License.
  */
 #include "velox/functions/lib/aggregates/SumAggregateBase.h"
-#include "velox/functions/prestosql/aggregates/AggregateNames.h"
 
 using namespace facebook::velox::functions::aggregate;
 
-namespace facebook::velox::aggregate::prestosql {
-namespace {
+namespace facebook::velox::functions::aggregate::sparksql {
 
-template <template <typename U, typename V, typename W> class T>
 exec::AggregateRegistrationResult registerSum(const std::string& name) {
   std::vector<std::shared_ptr<exec::AggregateFunctionSignature>> signatures{
       exec::AggregateFunctionSignatureBuilder()
@@ -33,13 +30,6 @@ exec::AggregateRegistrationResult registerSum(const std::string& name) {
           .returnType("double")
           .intermediateType("double")
           .argumentType("double")
-          .build(),
-      exec::AggregateFunctionSignatureBuilder()
-          .integerVariable("a_precision")
-          .integerVariable("a_scale")
-          .argumentType("DECIMAL(a_precision, a_scale)")
-          .intermediateType("VARBINARY")
-          .returnType("DECIMAL(38, a_scale)")
           .build(),
   };
 
@@ -64,40 +54,38 @@ exec::AggregateRegistrationResult registerSum(const std::string& name) {
         auto inputType = argTypes[0];
         switch (inputType->kind()) {
           case TypeKind::TINYINT:
-            return std::make_unique<T<int8_t, int64_t, int64_t>>(BIGINT());
+            return std::make_unique<SumAggregateBase<int8_t, int64_t, int64_t>>(
+                BIGINT());
           case TypeKind::SMALLINT:
-            return std::make_unique<T<int16_t, int64_t, int64_t>>(BIGINT());
+            return std::make_unique<
+                SumAggregateBase<int16_t, int64_t, int64_t>>(BIGINT());
           case TypeKind::INTEGER:
-            return std::make_unique<T<int32_t, int64_t, int64_t>>(BIGINT());
+            return std::make_unique<
+                SumAggregateBase<int32_t, int64_t, int64_t>>(BIGINT());
           case TypeKind::BIGINT: {
             if (inputType->isShortDecimal()) {
-              return std::make_unique<DecimalSumAggregate<int64_t>>(resultType);
+              VELOX_NYI();
             }
-            return std::make_unique<T<int64_t, int64_t, int64_t>>(BIGINT());
+            return std::make_unique<
+                SumAggregateBase<int64_t, int64_t, int64_t>>(BIGINT());
           }
           case TypeKind::HUGEINT: {
-            if (inputType->isLongDecimal()) {
-              return std::make_unique<DecimalSumAggregate<int128_t>>(
-                  resultType);
-            }
             VELOX_NYI();
           }
           case TypeKind::REAL:
             if (resultType->kind() == TypeKind::REAL) {
-              return std::make_unique<T<float, double, float>>(resultType);
+              return std::make_unique<SumAggregateBase<float, double, float>>(
+                  resultType);
             }
-            return std::make_unique<T<float, double, double>>(DOUBLE());
+            return std::make_unique<SumAggregateBase<float, double, double>>(
+                DOUBLE());
           case TypeKind::DOUBLE:
             if (resultType->kind() == TypeKind::REAL) {
-              return std::make_unique<T<double, double, float>>(resultType);
+              return std::make_unique<SumAggregateBase<double, double, float>>(
+                  resultType);
             }
-            return std::make_unique<T<double, double, double>>(DOUBLE());
-          case TypeKind::VARBINARY:
-            // Always use int128_t template for Varbinary as the result
-            // type is either int128_t or
-            // UnscaledLongDecimalWithOverflowState.
-            return std::make_unique<DecimalSumAggregate<int128_t>>(resultType);
-
+            return std::make_unique<SumAggregateBase<double, double, double>>(
+                DOUBLE());
           default:
             VELOX_CHECK(
                 false,
@@ -107,10 +95,5 @@ exec::AggregateRegistrationResult registerSum(const std::string& name) {
         }
       });
 }
-} // namespace
 
-void registerSumAggregate(const std::string& prefix) {
-  registerSum<SumAggregateBase>(prefix + kSum);
-}
-
-} // namespace facebook::velox::aggregate::prestosql
+} // namespace facebook::velox::functions::aggregate::sparksql
